+++ CHANGELOG.md
@@ -7,5 +7,6 @@ hi
 - All entities in the `minecraft:aquatic` tag can now swim through Algae without being slowed down.
 - Mesoglea and Nematocyst models are now data-generated.
 - Removed unused Nematocyst textures.
+- Shelf Fungi models are now data-generated.
 - Significantly optimized Sculk Sensor rendering with the `Minecraft Live Tendrils` resource pack.
 - Fixed many recipe unlocks being broken in 1.21.2+.